refactor : 패키지 분리 -> utils, message 분리
utils 클래스와 message 클래스 별도 패키지로 분리
