+++ src/main/java/baseball/message/Message.java
@@ -1,6 +1,6 @@
-package baseball;
+package baseball.message;
 
-public class Message {
+public abstract class Message {
 
     public static final String START_MSG = "숫자 야구 게임을 시작합니다.";
     public static final String END_MSG = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
