+++ src/main/java/baseball/utils/CheckUtils.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.utils;
 
 import java.util.Arrays;
